Make bookmarkModule.utils.js and move msnry object and removeBookmarkImages()
