+++ src/popup/bookmarkModule.utils.js
@@ -1,0 +1,20 @@
+import { elements } from './base';
+import { removeChildNodes } from '../utils';
+
+const Masonry = require('masonry-layout');
+
+export const msnry = new Masonry(elements.gridBookmarks, {
+  // options
+  itemSelector: '.grid-item',
+  columnWidth: '.grid-item',
+  gutter: '.gutter-sizer',
+  percentPosition: true,
+  transitionDuration: '0',
+});
+
+export function removeBookmarkImages() {
+  const div = document.createElement('div');
+  div.classList.add('gutter-sizer');
+  removeChildNodes(elements.gridBookmarks);
+  elements.gridBookmarks.appendChild(div);
+}
